Move `Compiler`, `CompilerOptions` etc under `CodeGen`
